requirements.txt
alabaster==0.7.11
atomicwrites==1.1.5
attrs==18.1.0
Babel==2.6.0
cachetools==2.1.0
certifi==2018.4.16
chardet==3.0.4
colorama==0.3.9
dataframe==0.2.1.3
docutils==0.14
google-api-python-client==1.6.7
google-auth==1.4.2
google-auth-httplib2==0.0.3
httplib2==0.11.3
idna==2.7
imagesize==1.0.0
Jinja2==2.10
MarkupSafe==1.0
more-itertools==4.2.0
nose==1.3.7
numpy==1.14.3
oauth2client==4.1.2
packaging==17.1
pandas==0.23.0
pluggy==0.6.0
py==1.5.4
pyasn1==0.4.3
pyasn1-modules==0.2.1
Pygments==2.2.0
pyparsing==2.2.0
pytest==3.6.2
python-dateutil==2.7.3
pytz==2018.4
requests==2.19.1
rsa==3.4.2
scipy==1.1.0
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.7.5
sphinxcontrib-websupport==1.1.0
tabulate==0.8.2
uritemplate==3.0.0
urllib3==1.23
